+++ yash-env/CHANGELOG.md
@@ -21,6 +21,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- The definition of `system::ChildProcessTask` is updated so that the `Output`
+  type of the returned `Future` is now `std::convert::Infallible` instead of
+  `()`.
 - External dependency versions:
     - Rust 1.85.0 → 1.86.0